LayoutTests/http/tests/media/media-source/video-media-source-play.html

<!DOCTYPE html>
<html>
    <head>
        <script src="/media-resources/video-test.js"></script>
        <script src="/media/resources/media-source/webm/segment-info.js"></script>
        <script src="media-source.js"></script>
        <script>
            var segmentHelper = new MediaSourceTest.SegmentHelper(WebMSegmentInfo.testWebM);

            function onSourceOpen(event) 
            {
                segmentHelper.addSourceId();
                segmentHelper.appendInitSegment();
                segmentHelper.appendUntilEndOfStream(0);
            }

            function onLoad() 
            {
                findMediaElement();

                waitForEventAndFail('error');
                waitForEvent('loadstart');
                waitForEvent('webkitsourceopen');
                waitForEvent('playing');
                waitForEvent('webkitsourceended');
                waitForEventAndEnd('ended');

                video.addEventListener('webkitsourceopen', onSourceOpen);

                segmentHelper.init(video, function(success)
                {
                    if (!success) {
                        failTest("Failed to load segment data");
                        return;
                    }

                    MediaSourceTest.setSrcToMediaSourceTestURL(video);
                });
            }
        </script>
    </head>
    <body onload="onLoad()">
        <video autoplay> </video>
        <p>Tests normal playback case with MediaSource API</p>
    </body>
</html>
